+++ examples/cpp/checksatassuming.cpp
@@ -56,11 +56,7 @@ main()
   // (check-sat-assuming ((= s0 goal)))
   result = bitwuzla.check_sat({mk_term(Kind::EQUAL, {s0, goal})});
   std::cout << "Expect: unsat" << std::endl;
-  std::cout << "Bitwuzla: "
-            << (result == Result::SAT
-                    ? "sat"
-                    : (result == Result::UNSAT ? "unsat" : "unknown"))
-            << std::endl;
+  std::cout << "Bitwuzla: " << result << std::endl;
 
   // (assert (= s1 (ite (= o0 (_ sortbv1 1)) (bvadd s0 one) s0)))
   bitwuzla.assert_formula(mk_term(Kind::EQUAL,
@@ -73,27 +69,20 @@ main()
   // (check-sat-assuming ((= s1 goal)))
   result = bitwuzla.check_sat({mk_term(Kind::EQUAL, {s1, goal})});
   std::cout << "Expect: unsat" << std::endl;
-  std::cout << "Bitwuzla: "
-            << (result == Result::SAT
-                    ? "sat"
-                    : (result == Result::UNSAT ? "unsat" : "unknown"))
-            << std::endl;
+  std::cout << "Bitwuzla: " << result << std::endl;
 
   // (assert (= s2 (ite (= o1 (_ sortbv1 1)) (bvadd s1 one) s1)))
-  bitwuzla.assert_formula({mk_term(Kind::EQUAL,
-                                   {s2,
-                                    mk_term(Kind::ITE,
-                                            {mk_term(Kind::EQUAL, {o1, one1}),
-                                             mk_term(Kind::BV_ADD, {s1, one2}),
-                                             s1})})});
+  bitwuzla.assert_formula(mk_term(Kind::EQUAL,
+                                  {s2,
+                                   mk_term(Kind::ITE,
+                                           {mk_term(Kind::EQUAL, {o1, one1}),
+                                            mk_term(Kind::BV_ADD, {s1, one2}),
+                                            s1})}));
 
   // (check-sat-assuming ((= s2 goal)))
   result = bitwuzla.check_sat({mk_term(Kind::EQUAL, {s2, goal})});
   std::cout << "Expect: unsat" << std::endl;
-  std::cout << "Bitwuzla: "
-            << (result == Result::SAT
-                    ? "sat"
-                    : (result == Result::UNSAT ? "unsat" : "unknown"))
-            << std::endl;
+  std::cout << "Bitwuzla: " << result << std::endl;
+
   return 0;
 }